requirements.txt
# Core dependencies
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
requests>=2.28.0

# Optional: Langfuse for observability
langfuse>=2.0.0

# Development
pytest>=7.0.0
